22:40 06/06/2016 Bản gần xong

+++ BTLCongNgheWeb_Version2/Models/CardItem.cs
@@ -10,8 +10,8 @@ namespace BTLCongNgheWeb_Version2.Models
         public int id { get; set; }
         public string name { get; set; }
         public int so_luong { get; set; }
-        public double gia { get; set; }
-        public CardItem(int _id,string _name,int _soluong,double _gia)
+        public int gia { get; set; }
+        public CardItem(int _id,string _name,int _soluong,int _gia)
         {
             id = _id;
             name = _name;
